solve the issue with gameState not defined
